requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
motor
pymongo
python-dotenv
PyJWT
cryptography
